// src/app/interfaces/card.interfaces.ts
export type Attribute =
	| 'fire'
	| 'water'
	| 'wood'
	| 'light'
	| 'dark'
	| 'heart'
	| 'jammer'
	| 'poison'
	| 'mortalPoison'
	| 'bomb'
	| '';
export type Attributes = Attribute[];
export type Type =
	| 'redeemable'
	| 'enhance'
	| 'awaken'
	| 'machine'
	| 'devil'
	| 'attacker'
	| 'god'
	| 'dragon'
	| 'healer'
	| 'physical'
	| 'balanced'
	| 'evo';
export type Types = Type[];